// src/app/enums/eprovider-actions.enum.ts
export enum EProviderActions {
    create_provider = "create_provider",
    create_contact = "create_contact",
    view_contact = "view_contact",
    delete_provider = "delete_provider",
    delete_contact = "delete_contact",
    edit_provider = "edit_provider",
    edit_contact = "edit_contact",

}